- refactor chare condition table - refactor operator picker - refactor home - use more memo hook
